+++ json_serializable/test/generic_files/generic_class.wrapped.dart
@@ -44,3 +44,70 @@ class GenericClass<T extends num, S> {
           [S other1, U other2]) =>
       {'value': input};
 }
+
+@JsonSerializable()
+@_DurationMillisecondConverter()
+@_DurationListMillisecondConverter()
+class GenericClassWithConverter<T extends num, S> {
+  @_SimpleConverter()
+  Object fieldObject;
+
+  @_SimpleConverter()
+  dynamic fieldDynamic;
+
+  @_SimpleConverter()
+  int fieldInt;
+
+  @_SimpleConverter()
+  T fieldT;
+
+  @_SimpleConverter()
+  S fieldS;
+
+  Duration duration;
+
+  List<Duration> listDuration;
+
+  GenericClassWithConverter();
+
+  factory GenericClassWithConverter.fromJson(Map<String, dynamic> json) =>
+      _$GenericClassWithConverterFromJson<T, S>(json);
+
+  Map<String, dynamic> toJson() => _$GenericClassWithConverterToJson(this);
+}
+
+class _SimpleConverter<T> implements JsonConverter<T, Map<String, dynamic>> {
+  const _SimpleConverter();
+
+  @override
+  T fromJson(Map<String, dynamic> json) => json['value'] as T;
+
+  @override
+  Map<String, dynamic> toJson(T object) => {'value': object};
+}
+
+class _DurationMillisecondConverter implements JsonConverter<Duration, int> {
+  const _DurationMillisecondConverter();
+
+  const _DurationMillisecondConverter.named();
+
+  @override
+  Duration fromJson(int json) =>
+      json == null ? null : Duration(milliseconds: json);
+
+  @override
+  int toJson(Duration object) => object?.inMilliseconds;
+}
+
+class _DurationListMillisecondConverter
+    implements JsonConverter<List<Duration>, int> {
+  const _DurationListMillisecondConverter();
+
+  @override
+  List<Duration> fromJson(int json) => [Duration(milliseconds: json)];
+
+  @override
+  int toJson(List<Duration> object) => object?.fold<int>(0, (sum, obj) {
+        return sum + obj.inMilliseconds;
+      });
+}