+++ json_serializable/CHANGELOG.md
@@ -12,6 +12,8 @@
 
 ## 1.4.0
 
+* Added support for `JsonConvert` annotation on fields.
+
 * `type_helper.dart`
 
   * `TypeHelper` `serialize` and `deserialize` have return type `Object` instead